Migrate everything to TypeScript (#51)
### Summary

This PR contains a lot of bulk changes to the code, so I'm grouping them all into one PR so that other  development branches hopefully only have to rebase once.

#### Migration to TypeScript

First and foremost, this PR migrates all of the code in the repository to TypeScript, unifying the language across the website, crawler, and backend. The start of this work landed in #50, and this PR finishes the migration for the rest of the components, hooks, and utilities. Using TypeScript in a repository this size is something that I personally believe is a big improvement over Vanilla JS:

- Using types self-documents the code so that contributors can rapidly come up-to-speed on component props, function arguments/return types, and class fields/methods without having to manually determine the types
- Using TypeScript lets the compiler catch type errors earlier instead of at runtime. For example, I believe this PR will fix #40, #38, and #39 (though I'll have to double-check) due to compile-time errors that I had to fix while migrating the code
- TypeScript integrates well with ESLint (via https://github.com/typescript-eslint/typescript-eslint) to catch additional classes of lint errors that require type information (as opposed to just operating on the AST), such as ensuring that errors in promises are handled.

In addition to the bare migration, this PR also contains a few substantive changes to the code, which fall into a few categories:

- restructuring code to work better with TypeScript (such as replacing .filter().map() with .flatMap() so that the TypeScript compiler's narrowing functionality would work)
- making the code more defensive to ensure that unexpected unknown types don't end up breaking code that assumes the correctness of the static types

#### Addition of lint rules

This PR also modifies the lint rules to better leverage TypeScript. Additionally, I added a lint rule about import order to promote more consistency and better visually separate external, internal, and styling imports. Moreover, I changed a prettier config to always add trailing commas, allowing future diffs to be cleaner.

#### Start of better logging/error handling

Right now, the website doesn't gracefully handle errors and doesn't leverage Sentry as best as it can. To start working towards greener pastures in this area, I added a couple functions, `hardError` and `softError` that can be used to wrap errors and report them to Sentry. A follow-up task to this will be to add more usages of `softError` (which is completely safe) throughout the code in unexpected code paths. Another follow-up will be to start incorporating error boundaries in the app to isolate errors and report them to Sentry with contextual information (I also might refactor hardError to not report to Sentry and let it bubble up to an error boundary instead).

### Tests/rollout plan

I've tested every feature I can locally, so I don't anticipate this PR will break anything, but due to its size that's hard to guarantee. I'll make sure to keep an eye on Sentry & on new issues, but even then the lack of traffic over the next month or two probably means bugs could go unnoticed until the class schedule for Spring 2022 is released.

#### Safe to rollback?

Yes