Sort courses by time in <Map> (#59)
### Summary

This PR sorts courses by the start time of their first meeting when constructing the course data map inside of `<Map>`. This affects both `<MapView>` and `<DaySelection>`. Additionally, in order to coalesce common behavior, this PR also makes `<Map>` construct a data structure with enough information for both `<MapView>` and `<DaySelection>` instead of doing each separately.

Finally, this PR also includes a minor change to `<MapView>` that fixes a DOM nesting issue with a `div` element being a descendant of a `p` element.

### Motivation

Fixes #45

### Rollout plan

This should be safe for CI to deploy, and shouldn't break any existing behavior.

This PR is safe to rollback.