Add scrolling to <DaySelection>, minor style tweaks (#60)
### Summary

This PR mainly fixes scrolling in `<DaySelection>`, which before would just cut off if there were too many courses to display on the same screen at the same time. Towards these ends, this PR also includes changes to the greater map styles that control the tab's content root's height.

In addition, this PR changes the way padding is determined between classes inside `<DaySelection>` to be a bit cleaner; the only visual change is that there is now a bit of padding between the course ID and the course rows (which probably should have been there before).

### Motivation

Ensuring that courses in `<DaySelection>` can be viewed even when there is a long list. Also, minor fixes to styles to improve spacing.

### Rollout plan

This should be safe for CI to deploy, and shouldn't break any existing behavior.

This PR is safe to rollback.