Separate InformationModal; enable dark mode on modals (#71)
### Summary

This PR separates `<InformationModal>` into its own component, as the start of larger changes to `<App>` that will eventually land with changes to the core data loading/fetching in #69. It also enables dark mode for the modal, which requires hoisting the dynamic theme CSS class (either `.light` or `.dark`) from the `div.App` element to the HTML `body` element. This is because the [swal](https://sweetalert2.github.io/) modals get appended to the body outside of the normal React hierarchy, so the existing cascades from `App.light` and `App.dark` don't work on modals.

![dark-mode-modal](https://user-images.githubusercontent.com/26242455/132422491-da0309ce-6547-4599-b62b-19278d88fc75.png)

Additionally, this PR also moves a bit of functionality out of `<App>` into a couple of smaller, private sub-components, `<ThemeLoader>` and `<AppCSSRoot>`.

### Motivation

Improve visuals by making color schemes apply to more places, and pave the way to more changes that will land in #69 and beyond.