Split <Header> into presentation & logic components (#73)
### Summary

This PR splits `<Header>` into `<Header>` and `<HeaderDisplay>`, which handle logic and presentation of the header, respectively. In addition, it moves a couple callback functions out to `utils.tsx` to reduce the amount of code in the components and create clean modular boundaries.

### Motivation

Separating the header into the two components will allow it to have the display parts re-used for app skeletons that are rendered when data is loading/being fetched. This functionality will land in #69.