Rename TermContext to ScheduleContext, organize contexts (#78)
### Summary

This PR renames `TermContext` (and any related types/values) to `ScheduleContext`, including the inner `patchTermData` (now `patchSchedule`) function. It also moves all contexts into their own files, grouping the value types & context objects in a more natural way.

Finally, this PR also includes a minor change to a lint rule: `prettier/prettier` is now a warning instead of an error, so it won't break compilation.

### Motivation

The renames are in preparation of merging a larger PR that moves data storage to local storage and includes support for multiple schedules per term (which will then be exposed in the UI in a later PR).