Remove index pattern from hooks directory (#79)
### Summary

Removes `src/hooks/index.ts` and adjusts all imports accordingly.

### Motivation

There are few uses of `hooks`, and there are not many items in the directory, so the utility of having an index file is dubious. Removing this makes it clearer where items are being imported from, and removes a file that has to be constantly upkept to ensure it has all sibling files re-exported.