Add test support, CI jobs (#83)
### Summary

This PR adds an example usage of the testing framework that create-react-app provides, in addition to a couple libraries/lint rules to make testing easier. Finally, it includes 3 new GitHub actions to handle CI for linting, building, and test-running, respectively (these all take on the order of 1-2 minutes, so they shouldn't negatively affect waiting very much).

### Summary

Better code quality with unit tests, and more confidence when merging with CI checks.