Migrate to local storage, add data-layer support for schedule versions (#84)
### Summary

This PR includes some major changes to the data layer of the application, many of which build on the work done in #69. Specifically, this PR contains a major migration of the backing storage of user data from cookies to local storage (see Motivation for the reasoning). Local storage is now used to store:
- Core schedule data
- UI theme
- [`<InformationModal>`](https://github.com/gt-scheduler/website/blob/development/src/components/InformationModal/index.tsx) seen flag

Along with the migration to local storage, I created a more reliable framework for implementing, running, and testing **migrations** on the schema of the stored schedule data (borrowing principles from the more general practice of [schema migrations from the database world](https://en.wikipedia.org/wiki/Schema_migration)). Whether to run a migration is determined by the `version` field, which identifies the schema version for the data. Then, by judiciously using TypeScript, the framework ensures that we migrate data to the latest schema before accessing it in the rest of the application (re-using the work done in #69 to wait to render the app until migration is done). To be more confident in the safety of these changes, I also added a comprehensive test suite for the core hooks/migration logic involved that includes mocks for the cookies and local storage APIs.

Finally, this PR includes a few other smaller but related changes:
- Core schedule data is now stored in a **single data key** (`schedule-data`). To make modifying this state tree easier, I introduced the usage of [Immer](https://immerjs.github.io/immer/) to make callbacks that can update the state using so-called "drafts" (copy-on-write mutable handles that perform immutable updates on the backing data once committed)
- In the first version of the new versioned schema, I also included support for including multiple schedule versions (though nothing in the UI surfaces this capability yet). My plan is to merge this and then rebase #23 to add the existing UI built there on top of the new data-layer capabilities.
- I formalized the pattern introduced in #69 where we have a series of components that render their children conditionally based on asynchronous data fetching state. Now, each of these components are called "stages", and they are orchestrated in the `<AppDataLayer>` component, which handles running all stages before providing the valid context values to their corresponding providers.

### Motivation

#### Moving to local storage from cookies

By migrating from storing user data in cookies to local storage, we gain a few key benefits:
- Effectively remove the cap on the size of the data stored. Cookies have a [maximum size of 4 KiB](https://docs.microsoft.com/en-us/previous-versions/ms178194(v=vs.140)?redirectedfrom=MSDN#cookie-limitations), whereas local storage can be at least 5 MB (a limit that schedule data will almost certainly never hit).
- Because the size limitation is now removed, this also opens up the opportunity to consolidate all schedule data into a single state tree, making update logic simpler and opening up the opportunity to more easily sync the schedule data in the future (#27)
- Cookies are also limited in *how many of them* we can have per-site, [with most browsers implementing a limit around 20](https://docs.microsoft.com/en-us/previous-versions/ms178194(v=vs.140)?redirectedfrom=MSDN#cookie-limitations). The scheduler was already getting dangerously close to this limit (for example, I have 15 cookies for gt-scheduler.org on my own computer), and the addition of schedule versions would likely cause it to start hitting the limit.
- Using cookies adds a small slow-down to the site, since the cookies are tacked on to every request the browser makes to the gt-scheduler.org domain. Unfortunately, this PR doesn't actually remove the existing cookies, but it will stop new users from incurring the slow-down.
  - The plan is to roll out these changes, and then in a patch a few months down the line (once registration for the next term starts and this new storage framework is battle tested), remove all existing data-related cookies (to remove their overhead).

#### Introduction of migration framework

The motivation behind the migrations is ultimately to:
- increase our confidence in the validity of the data stored in users' browsers (since we don't have access to it),
- to make it much easier to confidently change the schema in the future to support new features. The combination of TypeScript and testing lets us greatly increase our confidence in any future changes, and my hope is that it makes for a stronger codebase.

#### Adding data-layer support for schedule versions

These changes make it much easier to merge #23, which has unfortunately gone unmerged for too long (which is entirely my fault).