Move skeleton props into explicit object props on Stages (#101)
### Summary

This PR extends the existing `<Stage*>` component pattern to include passing a `skeletonProps` prop that is passed through the `<AppSkeleton>` component when the stage is loading. This decouples the breadth of the interactivity a stage will support while it is loading from the implementation, and instead moves it up to the usage of the stages (in `<AppDataLoader>`). This makes it easier to re-order stages.