Split ensure valid term to own stage (#102)
### Summary

This PR moves functionality out of `StageExtractTermScheduleData` (and its corresponding hook); namely, the check that makes sure the currently selected term is valid (in the list of terms downloaded from the GitHub API) before passing it down to the rest of the app. Factoring this functionality out to its own stage/hook makes it easier to see at a glance that the invariant is maintained.