Add DropdownMenu, move export actions to their own menu (#108)
### Summary

This PR adds a `<DropdownMenu>` component that is similar to a `<Select>` but is less complex. With this, the export (download schedule image, download ICS export, copy CRNs to clipboard) actions were all put into a single menu.

### Motivation

`<DropdownMenu>` is used in #93 to implement the account dropdown, and the export actions need to be in their own menu to make space on the header.