Fix local course GPA cache to skip storing empty data (#119)
### Summary

This PR fixes a bug added with #118 where course data would get stored (as empty data (`{}`) even if an error occurred making the request or parsing the data. The original PR intended to explicitly avoid this pit-fall, but ended up failing to do so. To fix this, the local storage key also needs to be changed to discard the old cached values.

This PR won't fix course GPAs not appearing on the scheduler; this requires re-gaining access to the Course Critique API.