Fix Mapbox Zoom In/Out (#209)
### Summary

Resolves #179 

This PR fixes the bug where zooming in and out on the Map using a
trackpad or mouse wheel is extremely slow and laggy. I upgraded
`react-map-gl` from `v5.2.10` to `v5.2.11` which didn't have this
performance issue. I did not upgrade to the latest version since it uses
MapBox v2 which hosts compatibility issues with Babel and would require
hacky fixes.

### Checklist

- [x] Zoom with a trackpad is not finicky
- [x] Zoom with a mouse is not finicky

### How to Test
- Upgrade module with `yarn install`
- Try out the map

---------

Co-authored-by: EmilyAL001 <eamanda21@gmail.com>
Co-authored-by: Hailey Ho <haileyho@BMESNEL-MBP14-A-2.local>