Recurring Events Export (#221)
### Summary

Resolves: N/A

With the addition of recurring events, we want users to be able to
export their events as well into a `.ics` calendar or image. However,
having only recurring events in the schedule does not enable this
feature. Moreso, the exported `.ics` calendar does not contain recurring
events. This PR fixes these issues and other minor bugs.

### Checklist

- [x] Fixed export not enabling with only recurring events
- [x] Fixed export to calendar not exporting recurring events
- [x] Fixed not being able to set start/end time between 9 PM and 10 PM

### How to Test
- Add events and try exporting to calendar or image

---------

Co-authored-by: Hailey Ho <haileyho@BMESNEL-MBP14-A-2.local>