Implement Drag and Drop for Custom Events (#212)
### Summary

Resolves #176 

Enabled drag and drop functionality for events. Users can change day and
time of events with a single day and only the time of events with
multiple days by dragging them.

NOTE: Events can't be selected on desktop screens anymore, but can do
further work to make drag and drop work with select functionality as
well if required.

### Checklist
- [x] User can click on the an `EventBlock` and drag it around.
- [x] While the `EventBlock` is being dragged, the block at the original
time should be "blurred"
- [x] Wherever the user drop the `EventBlock`, the event's time info
should be updated in `ScheduleContext`.
- [x] UI works for both light and dark modes.

### How to Test
Add an event and try dragging it.