Share schedule back feature (#289)
### Summary

Resolves #284 

added functionality to the UI for sharing back schedules to friends that
the user has schedules of

### Checklist

- [x] Extract the reference passed to the input div of the modal and
store in the component which will trigger the modal instead. Whenever
the user clicks share back, we will pass in the email of the friend that
they're trying to share their schedule with to the modal using the ref
that we extracted. This is a suggestion, feel free to use other
solutions which fulfill this goal. When the invitation modal opens, it
will have the email of the friend already input in the box and they will
just have to press send.
- [x] Create the UI elements as displayed in the figma prototype.
However instead of only having one of the share back options at a time
as shown in the pictures above, have both of them displayed at all times
for now.


### How to Test
Open compare schedules with friends not shared with

---------

Co-authored-by: Yatharth Bhargava <yatharthbhargava@gmail.com>