Brandon/293 share schedule UI changes (#296)
### Summary

Resolves #293

Edit style of expand button for comparison panel and Trigger special
LoginModal when user clicks 'share schedule' button or toggle compare
button without logging in

### Checklist

- [x] Edit styling of the expand button for the comparison panel (refer
to figma). Code for the button
- [x] Trigger LoginModal (in components folder) when user clicks 'share
schedule' button or toggle compare button without logging in. Look at
HeaderActionBar for details on how to use the LoginModal component.
Match the styling of the LoginModal to the styling of the figma and
change text as well (previous LoginModal usage shouldn't change so maybe
create props to determine text/styling).


### How to Test
attempt sharing while not logged in

---------

Co-authored-by: Yatharth Bhargava <yatharthbhargava@gmail.com>