Invitation modal (#287)
### Summary

Resolves #281

### Checklist

- [x] Modify InvitationModalContent in ``src/components/InvitationModal`
to now also use all of the version names from the ScheduleContext and
create the input component for selecting which versions to share.
- [x] Modify the request made to /createFriendInvitation to now use a
list of versions instead of a single version
- [x] When the modal opens, pre-select the current version the user is
on. Generate a new link and attach to copy link button, upon clicking
which the link is copied to the user's clipboard.
- [x] Everytime a user changes the selected versions, generate a new
link and attach to button.
- [x] Modify the logic for checking if the invitee has already accepted
an invite from the sender.
- [x] The ScheduleContext currently contains the invites sent only for
the version the user is currently on. The context needs to be updated to
now contain invites from all versions (can probably be done using
termScheduleData passed to the context provider).
- [x] Update the middle section of the modal accordingly and show
invites from all version instead of just current. Make this component
scrollable in case it becomes longer than a specified height.
- [x] Modify logic for removing friends now that there are multiple
schedule versions
- [x] Create link expiration dropdown and connect selection to
createLink function

### How to Test
Change cloud function base url to test

### Notes:
1. When I try to send email invite, I'm getting the error Error: connect
ECONNREFUSED 127.0.0.1:465 from nodemailer, I don't think this is my
code's problem? or maybe it is
2. Had to make couple changes to firebase functions, should I make pr to
firebase repo?
3. Question: "Modify the logic for checking if the invitee has already
accepted an invite from the sender." what are we doing if the invitee
has accepted invites for some but not all selected schedules?
4. How are we setting validFor field for links that never expire? put it
as 1000 days for now

---------

Co-authored-by: Yatharth Bhargava <yatharthbhargava@gmail.com>