Added Min Travel Time Option (#375)
### Summary

Resolves #364

<!-- What does this PR change and why? Discuss any breaking changes. -->
Adds min travel time option to schedule combination dropdown.

- Used MapBox API to generate distance array.
- `GT_LOCATIONS` array has some duplicate/different names for locations.
- Uses Oliver's PR's code for consistency.

### Checklist

- [x] Copy over the latitudes longitudes used in Ritwic's ticket to
`src/mapConstants`
- [x] Use the [MapBox Matrix
API](https://docs.mapbox.com/api/navigation/matrix/) once to get a
matrix of distances between all the possible class locations, store this
in the `src/mapConsts` file
- [x] Create a new sorting option in
[`src/data/beans/Oscar`](https://github.com/gt-scheduler/website/blob/594219a8b0985d9d79071558ab737e0161cbf1cd/src/data/beans/Oscar.ts#L177)
to sort by the sum of time taken to go from each location to the next
location on a particular day. This should automatically add it to the
combination container without you having to specify it.


### How to Test
<!-- Describe how to test your code. -->
Tested by starting a development instance with `npm start` and using the
option to create schedules. This assumes the MapBox distances are 100%
accurate.