Add Plan to Take metric (#376)
### Summary

Resolves #356 
- Modified the Oscar beans class to retrieve planned counts and store
it. I had to modify the constructor setup for this, because we want the
planned counts to be fetched before creating the Course beans.
- Modified Course and Section beans classes to contain field
`plannedCount` that stores the planned count for that course or section

### Checklist

- [x] Add a method(s) in the beans/ classes to retrieve the the
aggregate count information.
- [x] Update Course and Section types to contain this metric.


### How to Test
Deployed new cloud function to dev db and ran website locally. I used
console.log() to print the counts assigned to each section/course and
made sure it matched the count calculated by the function.

---------

Co-authored-by: Alexa Fazio <45372449+afazio1@users.noreply.github.com>